+++ src/Controller/ContactAndEventTransferController.php
@@ -11,6 +11,7 @@ use SALESmanago\Services\ContactAndEventTransferService;
 use SALESmanago\Entity\Contact\Contact;
 use SALESmanago\Entity\Event\Event;
 use SALESmanago\Services\SynchronizationService as SyncService;
+use SALESmanago\Services\CheckIfIgnoredService as IgnoreService;
 
 class ContactAndEventTransferController
 {
@@ -22,38 +23,86 @@ class ContactAndEventTransferController
      */
     protected $syncService;
 
+    /**
+     * @var IgnoreService
+     */
+    protected $ignoreService;
+
+    /**
+     * ContactAndEventTransferController constructor.
+     * @param Configuration $settings
+     */
     public function __construct(Configuration $settings)
     {
-        $this->settings = $settings;
-        $this->service  = new ContactAndEventTransferService($this->settings);
-        $this->syncService = new SyncService($this->settings);
+        $this->settings      = $settings;
+        $this->service       = new ContactAndEventTransferService($this->settings);
+        $this->syncService   = new SyncService($this->settings);
+        $this->ignoreService = new IgnoreService($this->settings);
     }
 
+    /**
+     * @param Contact $Contact
+     * @param Event $Event
+     * @return array
+     * @throws Exception
+     */
     public function transferBoth(Contact $Contact, Event $Event)
     {
+        if($this->ignoreService->isContactIgnored($Contact)) {
+            return array_merge(
+                $this->ignoreService->getDeclineResponse(),
+                ['settings' => $this->settings]
+            );
+        }
         return array_merge(
             [
                 'settings' =>
-                    $this->settings->setRequireSyncronization($this->syncService->isNeedSyncContactEmailStatus($Contact))
+                    $this->settings->setRequireSynchronization(
+                        $this->syncService->isNeedSyncContactEmailStatus(clone $Contact)
+                    )
             ],
             $this->service->transferBoth($Contact, $Event)
         );
     }
 
+    /**
+     * @param Event $Event
+     * @return array
+     * @throws Exception
+     */
     public function transferEvent(Event $Event)
     {
-        return $this->service->transferEvent($Event);
+        return array_merge(
+            [
+                Configuration::COOKIE_TTL => $this->settings->getCookieTtl()
+            ],
+            $this->service->transferEvent($Event)
+        );
     }
 
+    /**
+     * @param Contact $Contact
+     * @return array
+     * @throws Exception
+     */
     public function transferContact(Contact $Contact)
     {
-       return array_merge(
-           [
-               'settings' =>
-                   $this->settings->setRequireSyncronization($this->syncService->isNeedSyncContactEmailStatus($Contact))
-           ],
-           $this->service->transferContact($Contact)
-       );
+        if($this->ignoreService->isContactIgnored($Contact)) {
+            return array_merge(
+                $this->ignoreService->getDeclineResponse(),
+                ['settings' => $this->settings]
+            );
+        }
+
+        return array_merge(
+            [
+                'settings' =>
+                    $this->settings->setRequireSynchronization(
+                        $this->syncService->isNeedSyncContactEmailStatus(clone $Contact)
+                    )
+            ],
+            $this->service->transferContact($Contact)
+        );
     }
 
 }