+++ src/Entity/Contact/Contact.php
@@ -5,7 +5,6 @@ namespace SALESmanago\Entity\Contact;
 use SALESmanago\Entity\AbstractEntity;
 use SALESmanago\Exception\Exception;
 
-use SALESmanago\Entity\Contact\ApiDoubleOptIn;
 use SALESmanago\Helper\EntityDataHelper;
 
 class Contact extends AbstractEntity
@@ -24,23 +23,35 @@ class Contact extends AbstractEntity
         EXT_ID   = 'externalId',
         COOKIE_NAME = 'smclient';
 
-    private $email      = null;
-    private $contactId  = null;
-    private $fax        = null;
-    private $name       = null;
-    private $phone      = null;
-    private $company    = null;
-    private $state      = null;
-    private $birthday   = null;
+    private $email         = null;
+    private $contactId     = null;
+    private $fax           = null;
+    private $name          = null;
+    private $phone         = null;
+    private $company       = null;
+    private $state         = null;
+    private $birthday      = null;
     private $birthdayDay   = null;
     private $birthdayMonth = null;
     private $birthdayYear  = null;
-    private $externalId = null;
+    private $externalId    = null;
 
-    private $score = null;
+    private $score         = null;
 
-    private $Address = null;
-    private $Options = null;
+    /**
+     * @var Address|null
+     */
+    private $Address       = null;
+
+    /**
+     * @var Properties|null
+     */
+    private $Properties    = null;
+
+    /**
+     * @var Options|null
+     */
+    private $Options       = null;
 
     /**
      * @param array $contactData;
@@ -281,6 +292,27 @@ class Contact extends AbstractEntity
     	return isset($this->Options)
 		    ? $this->Options
 		    : $this->Options = new Options($options);
+    }
+
+    /**
+     * @param \SALESmanago\Entity\Contact\Properties $Properties
+     * @return $this
+     */
+    public function setProperties(Properties $Properties)
+    {
+        $this->Properties = $Properties;
+        return $this;
+    }
+
+    /**
+     * @param array $properties
+     * @return \SALESmanago\Entity\Contact\Properties
+     */
+    public function getProperties($properties = [])
+    {
+        return isset($this->Properties)
+            ? $this->Properties
+            : $this->Properties = new Properties($properties);
     }
 
     /**