+++ src/Model/ContactModel.php
@@ -6,7 +6,8 @@ namespace SALESmanago\Model;
 use SALESmanago\Entity\Contact\Contact;
 use SALESmanago\Entity\Contact\Address;
 use SALESmanago\Entity\Contact\Options;
-use SALESmanago\Entity\Contact\ApiDoubleOptIn;
+use SALESmanago\Entity\Contact\Properties;
+use SALESmanago\Entity\ApiDoubleOptIn;
 
 use SALESmanago\Entity\Configuration as Settings;
 use SALESmanago\Exception\Exception;
@@ -31,8 +32,9 @@ class ContactModel
      */
     public function getContactForUnionTransfer()
     {
-        $Address = $this->Contact->getAddress();
-        $Options = $this->Contact->getOptions();
+        $Address    = $this->Contact->getAddress();
+        $Options    = $this->Contact->getOptions();
+        $Properties = $this->Contact->getProperties();
 
         $contactRequestArray = [
             Settings::CLIENT_ID => $this->Settings->getClientId(),
@@ -43,6 +45,7 @@ class ContactModel
                 Contact::NAME    => $this->Contact->getName(),
                 Contact::PHONE   => $this->Contact->getPhone(),
                 Contact::COMPANY => $this->Contact->getCompany(),
+                Contact::EXT_ID  => $this->Contact->getExternalId(),
                 Contact::STATE   => $this->Contact->getState(),
                 Contact::ADDRESS => [
                     Address::STREET_AD => $Address->getStreetAddress(),
@@ -51,20 +54,22 @@ class ContactModel
                     Address::COUNTRY   => $Address->getCountry()
                 ],
             ],
-            Settings::OWNER      => $this->Settings->getOwner(),
-            Options::N_EMAIL     => $Options->getNewEmail(),
-            Options::F_OPT_IN    => $Options->getForceOptIn(),
-            Options::F_OPT_OUT   => $Options->getForceOptOut(),
-            Options::F_P_OPT_IN  => $Options->getForcePhoneOptIn(),
-            Options::F_P_OPT_OUT => $Options->getForcePhoneOptOut(),
-            Options::TAGS        => $Options->getTags(),
-            Options::R_TAGS      => $Options->getRemoveTags(),
-            Contact::BIRTHDAY    => $this->Contact->getBirthday(),// attention
-            Address::PROVINCE    => $Address->getProvince(),// attention
-            Options::LANG        => $Options->getLang()
+            Settings::OWNER        => $this->Settings->getOwner(),
+            Options::N_EMAIL       => $Options->getNewEmail(),
+            Options::F_OPT_IN      => $Options->getForceOptIn(),
+            Options::F_OPT_OUT     => $Options->getForceOptOut(),
+            Options::F_P_OPT_IN    => $Options->getForcePhoneOptIn(),
+            Options::F_P_OPT_OUT   => $Options->getForcePhoneOptOut(),
+            Options::TAGS_SCORING  => $Options->getTagScoring(),
+            Options::TAGS          => $Options->getTags(),
+            Options::R_TAGS        => $Options->getRemoveTags(),
+            Contact::BIRTHDAY      => $this->Contact->getBirthday(),// attention
+            Address::PROVINCE      => $Address->getProvince(),// attention
+            Options::LANG          => $Options->getLang(),
+            Properties::PROPERTIES => $Properties->get()
         ];
 
-        if ($this->isSubscribtionStatusNoChangeChecker()) {
+        if ($this->isSubscriptionStatusNoChangeChecker()) {
             $contactRequestArray[Options::F_OPT_IN]              = false;
             $contactRequestArray[Options::F_OPT_OUT]             = false;
             $contactRequestArray[Options::F_P_OPT_IN]            = false;
@@ -72,15 +77,30 @@ class ContactModel
             $contactRequestArray[ApiDoubleOptIn::U_API_D_OPT_IN] = false;
         }
 
+        if ($this->Contact->getOptions()->getIsUnSubscribes()) {
+            $contactRequestArray[Options::F_OPT_IN]    = false;
+            $contactRequestArray[Options::F_OPT_OUT]   = true;
+            $contactRequestArray[Options::F_P_OPT_IN]  = false;
+            $contactRequestArray[Options::F_P_OPT_OUT] = true;
+        }
+
+        if ($this->Contact->getOptions()->getIsSubscribes()) {
+            $contactRequestArray[Options::F_OPT_IN]    = true;
+            $contactRequestArray[Options::F_OPT_OUT]   = false;
+            $contactRequestArray[Options::F_P_OPT_IN]  = true;
+            $contactRequestArray[Options::F_P_OPT_OUT] = false;
+        }
+
         if ($this->apiDoubleOptInChecker()) {
             $ApiDoubleOptIn = $this->Settings->getApiDoubleOptIn();
 
-            $contactRequestArray = array_merge([
+            $contactRequestArray = array_merge(
+                $contactRequestArray, [
                 ApiDoubleOptIn::U_API_D_OPT_IN        => $ApiDoubleOptIn->getEnabled(),
                 ApiDoubleOptIn::D_OPT_IN_EMAIL_ACC_ID => $ApiDoubleOptIn->getAccountId(),
                 ApiDoubleOptIn::D_OPT_IN_TEMPLATE_ID  => $ApiDoubleOptIn->getTemplateId(),
                 ApiDoubleOptIn::D_OPT_IN_EMAIL_SUBJ   => $ApiDoubleOptIn->getSubject()
-            ], $contactRequestArray);
+            ]);
         }
 
         return DataHelper::filterDataArray($contactRequestArray);
@@ -102,48 +122,52 @@ class ContactModel
     }
 
     /**
-     * Create Contact from Basic responce;
-     * @param $response
-     * @retrun Contact
+     * Create Contact from Basic response;
+     * @param array $response
+     * @return Contact $Contact
      */
-    public function getContactFromBasicResponse($responseContact)
+    public function getContactFromBasicResponse($response)
     {
         $this->Contact
-            ->setName($responseContact['name'])
-            ->setEmail($responseContact['email'])
-            ->setPhone($responseContact['phone'])
-            ->setFax($responseContact['fax'])
-            ->setScore($responseContact['score'])
-            ->setState($responseContact['state'])
-            ->setCompany($responseContact['company'])
-            ->setExternalId($responseContact['externalId'])
-            ->setContactId($responseContact['contactId'])
-            ->setBirthdayYear($responseContact['birthdayYear'])
-            ->setBirthdayMonth($responseContact['birthdayMonth'])
-            ->setBirthdayDay($responseContact['birthdayDay']);
+            ->setName($response['name'])
+            ->setEmail($response['email'])
+            ->setPhone($response['phone'])
+            ->setFax($response['fax'])
+            ->setScore($response['score'])
+            ->setState($response['state'])
+            ->setCompany($response['company'])
+            ->setExternalId($response['externalId'])
+            ->setContactId($response['contactId'])
+            ->setBirthdayYear($response['birthdayYear'])
+            ->setBirthdayMonth($response['birthdayMonth'])
+            ->setBirthdayDay($response['birthdayDay']);
 
         $this->Contact->getOptions()
-            ->setOptedOut($responseContact['optedOut'])
-            ->setOptedOutPhone($responseContact['optedOutPhone'])
-            ->setDeleted($responseContact['deleted'])
-            ->setInvalid($responseContact['invalid'])
-            ->setModifiedOn($responseContact['modifiedOn'])
-            ->setCreatedOn($responseContact['createdOn'])
-            ->setLastVisit($responseContact['lastVisit']);
+            ->setOptedOut($response['optedOut'])
+            ->setOptedOutPhone($response['optedOutPhone'])
+            ->setDeleted($response['deleted'])
+            ->setInvalid($response['invalid'])
+            ->setModifiedOn($response['modifiedOn'])
+            ->setCreatedOn($response['createdOn'])
+            ->setLastVisit($response['lastVisit']);
 
         $this->Contact->getAddress()
-            ->setStreetAddress($responseContact['address']['streetAddress'])
-            ->setZipCode($responseContact['address']['zipCode'])
-            ->setCity($responseContact['address']['city'])
-            ->setCountry($responseContact['address']['country']);
+            ->setStreetAddress($response['address']['streetAddress'])
+            ->setZipCode($response['address']['zipCode'])
+            ->setCity($response['address']['city'])
+            ->setCountry($response['address']['country']);
 
         return $this->Contact;
     }
 
-    protected function isSubscribtionStatusNoChangeChecker()
+    /**
+     * Checks if subscription status no change
+     * @return bool
+     */
+    protected function isSubscriptionStatusNoChangeChecker()
     {
         return (!$this->Contact->getOptions()->getIsSubscribes()
-            && $this->Contact->getOptions()->getIsSubscribtionStatusNoChange()
+            && $this->Contact->getOptions()->getIsSubscriptionStatusNoChange()
             && !$this->Contact->getOptions()->getForceOptIn());
     }